+++ src/Contexts/Movement/Infrastructure/Persistence/MovementEntity.ts
@@ -1,10 +1,9 @@
-import { Entity, PrimaryGeneratedColumn, Column } from "typeorm";
+import { Entity, Column } from "typeorm";
 import { Movement } from "@Movement/Domain/Movement";
+import { BaseEntity } from "@Shared/Infrastructure/TypeOrm/Objects/BaseEntity";
 
 @Entity("tbl_movement")
-export class MovementEntity implements Movement {
-    @PrimaryGeneratedColumn()
-    id: number;
+export class MovementEntity extends BaseEntity implements Movement {
 
     @Column()
     name: string;
