+++ src/Contexts/Shared/Infrastructure/TypeOrm/Objects/BaseEntity.ts
@@ -1,0 +1,12 @@
+import { PrimaryGeneratedColumn, CreateDateColumn, UpdateDateColumn } from "typeorm";
+
+export abstract class BaseEntity {
+    @PrimaryGeneratedColumn()
+    id: number;
+
+    @CreateDateColumn()
+    created_at: Date;
+
+    @UpdateDateColumn()
+    updated_at: Date;
+}
